Updated openstack/openstack Project: openstack/neutron  07d0761ce9cbb54578c126d19612acfc0a953cf6
Elevate context for router lookups during floating IP association.

During a floating IP association, the tenant making the request
may not always be the owner of the router.  To make the association,
Neutron needs to query the router details internally but needs to
use an elevated context to do so.  Otherwise, the user sees a
cryptic error stating that the router doesn't exist.

Change-Id: If2bd6baa785ff879c61ce12e70b62e0ba25635f5
Closes-Bug: #1537754