Updated openstack/openstack Project: openstack/neutron  66f4e0dd1c85545eae6d51683f382e1f280e9203
Introduce new queries to return DVR routers for a host

When we remove explicit binding of dvr routers to compute nodes
we'll need a way to know which dvr routers should be returned for
a particular host in order to serve sync_routers requests from agents.
This patch adds such queries and updates l3 rpc handler to use it.

Partially implements blueprint improve-dvr-l3-agent-binding

Change-Id: I4ed4c0e50b7d261bbfd95e778a8fa2594f78e969