Updated openstack/openstack Project: openstack/designate  79a4af04556d904f9f4e4b9d9a91b54652b9f287
Make supported record type configurable

Also made the distinction between 'valid' and 'supported' record
types. For example an error of "'foo' is not a valid record type"
will return when user tries to create a record type of 'foo'.
User will get "'SPF' is not a supported record type" if operator
chooses not to expose SPF.

Implements blueprint configurable-record-type
Change-Id: I22410a760a5ec78270162496ee03f2b2b8ba7d25