Updated openstack/openstack Project: openstack-infra/shade  9a28d08e76328d7d533dbe821fe4e972067740ec
include keystonev2 role assignments

This will require that user is set with project possibly being None,
otherwise we would have to query over all combinations of users and
projects to find all the assignments, which I would say is unreasonable.

project is not required by keystoneclient, but based on what I was told
by someone in #openstack-keystone, it is highly discouraged to create
roles and users and assignments without the project being involved, so
we won't be allowing it.

Change-Id: Id5b7b9fb44a9dbecb7488eb8f0ef30773efed6d2