Updated openstack/openstack Project: openstack/heat  a7afe175b23f0aedecc2c4b8ac4145b53a73770b
Add timeout property for resolving attributes

When we resolve attributes of most part of resources we make a real call
to correspoding service API and wait a response.
This patch adds new property 'attr_wait_secs', which simulates delay
during resolving attributes.
It can be used for testing huge stacks with a lot of outputs and resources.

Change-Id: I996164b6e97bb86f561b984b6bb49af6883cedc1