Updated openstack/openstack Project: openstack/manila  b81b98b2b7455c4e9073243f86c7deda41c5784e
Fix db shim layer mismatches with implementation

The module manila/db/api.py is a pass-through shim
layer intended to provide an interface between the
implementation and the actual database queries implemented
elsewhere. However, many interface methods in this
module were not in sync with the actual implementation.

Remove unused interfaces and create a unit test to avoid
mismatches in the future.

Co-Authored-By: Goutham Pacha Ravi <gouthamr@netapp.com>

Change-Id: I275f028bf6f6afe10eb71fa005d3caefbc5673d2
Closes-Bug: #1516160