Updated openstack/openstack Project: openstack/heat  d012641693e5208d170120e9315046a8707e1e9a
Don't pass [] for fixed_ips to neutron

Earlier we used to remove `fixed_ips` attibute while calling neutron,
if  it's [] for both create and update. This was masking the
behaviour of releasing fixed_ips from a port during update or creating
a port without fixed_ips.

The recent change[1] to fix this is backward incompatible with some
tripleo templates.

This patch fixes it by reverting to the original behaviour.

[1] https://review.openstack.org/#/c/256328/

Change-Id: I72d9533ff7ed7656224849f6b134652db0532aef
Closes-Bug: #1538473