Updated openstack/openstack Project: openstack/neutron  8f8b07d396ef531a7a0f50aa3437c6c7674dd6cd
sriov-mech: Introduce a new VIF type for PF vnic type

We need to tell Nova to use a specific VIF type for full PF passthrough
(this is the recently added VNIC_DIRECT_PHYSICAL vnic type).

Make sure that when binding ports with this vnic type, we set the
appropriate vif type for Nova to be able to provision those ports
correctly.

Nova change that adds a vif driver for this vif type:

https://review.openstack.org/#/c/262583/

DocImpact: Exposes a new vif type for PF passtrhough vnic type

Change-Id: I895dab98e5e5a9369771539e51ba5c500bfe0045
Related-blueprint: sriov-pf-passthrough-neutron-port