Updated openstack/openstack Project: openstack/nova  1c8b92f1ee9ae08e23fe91d42ebc1d2f1a9eec7d
We now require gettext for dev environments

Post Ic1f106684270a8f2197e0fb2a67d6fc00535ee65 we now require
msgfmt (a binary in the gettext package) to be installed for pep8
to pass. This hasn't failed in the gate because infra already
installs that package. We should make sure new devs know to do
that thing as well.

Change-Id: Iac505ddd54466e8a6272fee5ff61aac180c90ac4