Updated openstack/openstack Project: openstack-dev/devstack  fff59d4fbb5ba938f5f4467f59c1a2370ba3dde6
Remove option 'DEFAULT/verbose'

Option "verbose" from group "DEFAULT" is deprecated for removal.
Its value may be silently ignored in the future.

If this option is not set explicitly, there is no such warning.
Furthermore, the default value of verbose is true, so there is
no need to set this value in config files.

TrivialFix

Change-Id: I581238e661c2eb0c29dd915d1b1e4773dcdeb4c8