Updated openstack/openstack Project: openstack/oslo.config  9f6d13e81b0de48b4590182102bf21a81823a1b0
Make _Namespace the value-holder

At present value-holding responsibilities are split between _Namespace
and MultiConfigParser. _Namespace holds the CLI options (as attrs)
while MultiConfigParser holds the rest. I for one find it odd that a
'parser' holds values. MultiConfigParser isn't really a parser at all
as it turns out, it just has a 'read' method bolted on that we don't
use (it is called by the networking_cisco project).

This patch takes the value-holding parts of MultiConfigParser and
duplicates them onto _Namespace so the latter no longer needs the
former. I have deprecated MultiConfigParser so as not to break
networking_cisco.

Change-Id: I87f4edce84685a53b2ca9096b595234dc549ec13