Updated openstack/openstack Project: openstack/oslo.config  ff71f06a66903d1f525113af6fa7f9d7cd616816
Make list_all_sections return a list

There's no real benefit to returning an iterator and now we have to
manage two namespaces it'd be a pain to return a deduped iter without
at least one traversal. The list won't be very long and the only usage
I can find is in Sahara, where they listify the iterator they receive.
So we turn a set into a sorted list into an iter into a list, pretty
pointless.

Change-Id: Ie8f9531a783b4ab6ffae1920d49a2966bc53eb4a