Updated openstack/openstack Project: openstack/diskimage-builder  53c57d75c786f7b4f1aa267c93d2c5825e232767
Cleanup unmount_dir function

Cleanup this function to work with a symlinked directory.  Document
it's behaviour more exactly, and add a simple unit-test for it (not
run by default, due to doing things like mounting and unmounting
system dirs on a live system, which doesn't seem safe for CI.  But it
is useful for developers ensuring sanity).

Change-Id: I335316019ef948758392b03e91f9869102a472b9