Updated openstack/openstack Project: openstack/heat  c864504117c2573d685e7cad47a941fbbdb1771e
Allow StackLock to be used as a context manager

We initially didn't implement this class as a context manager because it
was being used exclusively to lock operations that were happening in an
asycnronous thread (hence the thread_lock() context manager). However, we
now have some uses for it in the same thread (see
Ifa48b179723a2100fff548467db9e162bc669d13), so it makes sense to be able to
use the lock object itself as a context manager.

Change-Id: Ib90df8f8a149b57b0588d4fc9951f7f035ad7567