Updated openstack/openstack Project: openstack/neutron  f40538e6290dbf97276b1e7f4ec68c8cad0f9e47
create_object should not add an ID if not present in the DB model

Change I192793aa433606a1508f24564c54164f961018d9 introduced a way
to handle DB objects that don't have 'id' as primary field.
Anyway create_object should be modified not to add the 'id' field
if not present in the model.

Change-Id: I14e4e32255caf9244ff2793aef142b1a6fa57de3