Updated openstack/openstack Project: openstack-infra/storyboard  cb7943d945200b50eee1cbbd32493d9579bd4300
Move summarize_task_statuses into wmodels.Story

Having summarize_task_statuses in the stories database API meant that
the stories database API had to import wmodels. This makes it
impossible to use the stories API in wmodels.py, which is something
that will be useful for optimising worklists and boards.

Change-Id: Ieafadbe2029f29c14bfa54afe64f06bd29a79dd4