Updated openstack/openstack Project: openstack-infra/storyboard  a34d6a31292cf7b10fdd70cfb8484f555cf9cdeb
Resolve Board and Worklist contents on the server-side

This patch adds fields to the wsme models for boards and worklists
to represent the contents of each thing in a better way. This lets
us resolve the contents of a worklist or a board when handling the
initial GET request rather than resolving the contents by having
the webclient make a large number of additional GET requests.

This will speed up the board detail pages, and remove the issue
where each card appears one-at-a-time. The webclient will need to
be updated to understand the new representations in order to see
this speedup.

It also speeds up some parts of the boards and worklists APIs
by not needlessly looking up the board/worklist repeatedly.

NOTE: This patch preserves the existing functionality, so old
webclients will work fine.

Change-Id: Ie6c527ce87866c9a12d1a8f15ecf6e8cc319b71f