Updated openstack/openstack Project: openstack/ceilometer  501143062b82da180b0f9c40e4a505d833892e92
Adds timestamp option to Aggregation transformer

Adds an argument to the Aggregation transformer constructor
that allows a user to specify whether to include the timestamp
from either the first or last sample received for a given
aggregated sample.

This addresses an issue with transformer chaining where incorrect
values will sometimes be produced by the Rate of Change
transformer when chaining the Aggregation transformer with the
Rate of Change transformer.

Change-Id: Ib163a80a7e6ddaf58d7cc555fb4f4d87d570b1a1
Closes-Bug: #1539163