Updated openstack/openstack Project: openstack/nova  159f379e1688923c30fcad9d10e2a44c0d2e6a38
Adds json sample for the versioned notifications

This patch adds infrastructure to store and verify json samples
about versioned notifications. Also it adds sample for the service.update
notification as that is the only existing versioned notification.

A new notification_sample_tests subtree added to the functional test
tree to verify that the emitted notifications match with the stored sample.

Partially-Implements: bp versioned-notification-api
Change-Id: I3a70b7b5d6d427611a9d2c1610e0d55aff88f0e8