Updated openstack/openstack Project: openstack-infra/irc-meetings  ef40f1468099e9b82ec92be0f49bc083237b3260
Adjust api-wg meeting time to match reality

The api working group bi-weekly schedule has become out of sync with the
actual meeting times. This was most likely due to the time taken off
over the holiday break. This change corrects that inaccuracy and fills
in the "chair" field as well.

* swap the start_date fields for the biweekly meetings
* swap the frequency fields for the biweekly meetings
* add etoews, elmiko and cdent as chairs

Change-Id: Icf7f26fbcf9984f33f17e87e3e423f70458ee85b