Updated openstack/openstack Project: openstack/nova  2e9c6dc3f105161335bde3966b061c3c0e34457f
Adding guard on None value for some helpers method

We use some helper methods to strip and extrapolate details from the
device name string, some of these methods are missing the check on None
values, if a None is passed the code raise a TypeError exception. This
change adds a check for managing None values correctly.

Change-Id: Iebfaa09e494dac085b78dbaffa867f8047c2106b