Updated openstack/openstack Project: openstack/nova  b7461ffb5cb12d29fbd8834b0e3e95f2a0c932c4
libvirt: force use of direct vs p2p migration

If live_migration_flag or block_migration_flag does not include the
VIR_MIGRATE_PEER2PEER libvirt migration flag, then migrations will
fail with a warning like:

  Live Migration failure: Requested operation is not valid:
  direct migration is not supported by the connection driver

We should avoid invalid configurations like this, and warn users about
it.

However, there is a caveat - with the xen virt type, the reverse is
actually true. Peer-to-peer migrations are not supported in libvirt's
xen driver, so we must remove the VIR_MIGRATE_PEER2PEER for migrations
to work with that driver.

Change-Id: I7fa43bd2025d067181a86b27952f2cad428aa823
Signed-off-by: Mark McLoughlin <markmc@redhat.com>