Updated openstack/openstack Project: openstack/nova  c283518ecdb8a114a47250b5185c9b621f10ea87
Add better help text to scheduler options (3)

The previous commit moved all the scheduler options to a central
location. This commit adds additional help text to those options.
Because of the large number of options, the improved help text will be
split into several patches; this is patch 3 of 7.

Blueprint centralize-config-options

Change-Id: Idc4f94cf7943992e8beb4a81642a9a24afdf4d8b