Updated openstack/openstack Project: openstack/keystone  ff261669f404c48c7f9cf5ff800c6c8a22979da1
remove KVS backend for keystone.contrib.revoke

when the revoke extension was moved out to the top level, a
reference to already deprecated KVS backend was kept.

the reference in contrib was also deprecated, but since we're
removing the one it points to, there is no need to keep it
around.

bp removed-as-of-mitaka
Change-Id: Ib917b2daff70bd942a8f9786998ca0ed743c28d3