Updated openstack/openstack Project: openstack/keystone  afef20658a278322d93ff5e86798161b6b20eff1
Fix schema validation to use JSONSchema for empty entity

APIs such as region creation allow an empty entity (a random UUID will
be generated as the region in this case).

Keystone uses JSONShema for validation and will allow an empty entity
to be passed if the schema doesn't have `required` properties (as the
region creation schema allows).

The patch fixes the schema validation code to pass even an empty entity
to JSONSchema validation so that an API such as region create can pass
the validation and can be created successfully.

Change-Id: If3dd49af5e16bb3b741efa4573f9f8e9085ddd14
Closes-Bug: #1501740