Updated openstack/openstack Project: openstack/horizon  39959243486346a9a4ddb0ed39d870e1633344bf
Style: Login page should not have scroll bars

The login page is a bit longer than it should be, and as such, it is
scrollable.  This is not immediately apparent in OSX, as you have to
actually scroll in the window for the scrollbars to become visible,
but it was apparent in non-OSX operating systems.

We add a top padding to the application to compensate for the navbar
height, however, we don't need to add that padding to the splash
page because it doesn't have a navbar.

Change-Id: If41b0020dc2c4221ec2aa9fecb42422bddaf681d
Closes-bug: #1538301