Updated openstack/openstack Project: openstack-infra/project-config  c2df697e50a66e1ee9c2273b2e952a027780baff
Add windmill project

The purpose of windmill is to deploy the CI ansible roles. We'd be
using it as both a integration test, like devstack, and offer real
world examples on how to consume the ansible roles.

Windmill is currently an unofficial OpenStack project. For the
moment we won't be using IRC notification.

Change-Id: I16da59cf34db4bf949d93edd0f4359c47695ddf0
Signed-off-by: Paul Belanger <pabelanger@redhat.com>