Updated openstack/openstack Project: openstack/horizon  83113bf31ed7265f99521e81e4923f64f2d9e582
Fix metadata tree available filter

This patch updates the filtering of the available metadata items in
the Update Metadata modal so that any leaf with either a display
value or property name that matches will be displayed. Parent items
of any matching leaf are also displayed so context and the tree
structure are preserved.

Closes-Bug: #1519535
Closes-Bug: #1519567
Closes-Bug: #1463733
Change-Id: I11caf506a851807cb3b079184dd260db6862b3d9