Updated openstack/openstack Project: openstack/oslo.versionedobjects  8e5a40a977b3e02c007162fd3556d16dc5e61d7b
Add ability to pass args/kwargs to obj_class init

Because some subobjects of VersionedObject my require arguments to be
passed into __init__(), there needs to be a way for the
test_compatibility_routines() fixture to pass required args/kwargs to
init on certain objects. This change adds the ability to pass 2 mappings
to test_compatibility_routines, one for args and one for kwargs.

The key of each mapping is the object class that needs the special init
args. The value in the args dict will be a list of arguments to pass
(becomes *args), and the value in the kwargs dict is a list of kwargs
to pass (becomes **kwargs). If there is no key defined, the default
value of the args dict is [], and the default of kwargs is {}. *[] is
the equivalent of sending no arguments, and **{} is the equivalent of
not sending kwargs.

Change-Id: I62646b99adca47a9c9fe0f466f7a23ac8fa4553e
Related-Bug: #1537882