Updated openstack/openstack Project: openstack/neutron  344e97aadda835c8c28ce2ed31620527be201319
Add relationship between port and floating ip

We need to have the relationship between port and floating ip, because updating
quota will happen when the event "after_delete" occurs. And current cascade
removal of the floating ip does not cause the event "after_delete" for floating
ip. The cascade on the ORM-level "delete" must be added.

Closes-bug: #1533687
Change-Id: Id414cf20069563085737fee5ddce94b1bf2f912d