Updated openstack/openstack Project: openstack/horizon  03fab20aabba5f7f7f9d25b340908645ec30fb53
Fix errors on keyboard nav across tabs on forms

There is code in place to allow using the Tab / Shift+Tab keys to be
used for navigating across all tabs within a form. This fixes a bad
reference in this code and restores the intended behavior.

The problem occurs because the code assumes the implicit "event"
variable exists, but that is not always the case. Specifically this
seems to be a problem in Firefox. Instead of relying on the implicit
variable the event object should be referenced using the "evt"
parameter.

Closes-Bug: #1482353
Change-Id: I15812693a3d89c0862067119293004e97139040a