Updated openstack/openstack Project: openstack/cinder  6ac895e13b8e27c53ec81363f1ed2d37b54d6d0d
Add missing RPC calls versions to rpcapi modules

Some RPC calls in rpcapi modules are not passing version indicators to
the RPC client. The RPC call version should be indicated explicitely
because otherwise it gets defaulted to the latest one or '1.0'. This
creates a set of problem in maintaining RPC API backward compatibility.

This commit changes these calls to point to the right versions according
to git history and version comments in rpcapi files.

Change-Id: Ib05d9748e18589df5a67f7b7ea668394cc8df769
Related-Blueprint: rpc-object-compatibility