Updated openstack/openstack Project: openstack/nova  a2e11fc6b937073b13980362370e0140fc38ee08
Use extra_data_func to get fingerprints of objects

A kwarg was added to get_hashes in o.vo that allows extra data to be
gathered locally when generating the fingerprints. Since notifications
also need to be versioned on the internally held schema, this data is
added to the fingerprint data when the object inherits from the base
notification payload class.

The fingerprint generated for NotificationPayloadBase and
ServiceStatusPayload changed because of the new function, not because
of any added/removed data, so the version was not bumped.

Change-Id: Ic94c044d3a76a236ac4361b2681ae6da36d5b014
Partially-implements: bp versioned-notification-api