Updated openstack/openstack Project: openstack/nova  b2cc38d572a6431e4c37d5f37420781a334da7ef
Add better help text to scheduler options (4)

The previous commit moved all the scheduler options to a central
location. This commit adds additional help text to those options.
Because of the large number of options, the improved help text will be
split into several patches; this is patch 4 of 7.

Blueprint centralize-config-options

Change-Id: I0fdfb43e7be906f39888faadaa096782201afd88