Updated openstack/openstack Project: openstack/openstack-doc-tools  8e6127f057b24f5a5385ed0ed3814f60540c963b
Add switcher for common directory

To migrate common directory for rst files from common-rst to common,
Add a switcher for common directory as a temporary trick.
When find a rst file in common directory,
we use "common" directory as priority.
This switcher will be removed after all migration works finish.

Change-Id: I7ba517136807513f982e66010906b97621f844fc