Updated openstack/openstack Project: openstack/trove-integration  b5e7c20bd893b2abf816ea432437607de00fe5a0
Added enum to elements and admin_alt_demo to tests

The Datastore Log Operations have a dependency on ENUM, so it has
been added to the elements.

There is also a need for an 'admin' user in the alt_demo project,
in order to test the functionality fully.  This has been added
as well.

An argument was also added to the cmd_dsvm_gate_tests method to
facilitate running the scenario tests in the check/gate.

Partially implements: blueprint datastore-log-operations

Change-Id: Id449582954b0e3a7e95ef15418bdab3995cb7097