Updated openstack/openstack Project: openstack/nova  e43c67699de679af4f86bc0a8c8fece03bcdafa9
Ironic: Lightweight fetching of nodes

The Ironic API version >= 1.8 supports fetching only a small number of
a resource instead of it's full representation. This patch is pinning
the API version for Ironic at 1.8 so the nova driver can benefit from
this feature by fetching only the required fields for the deployment,
making it more lightweight and improving the readability of the logs.

The configuration option "api_version" was marked as deprecated and
should be removed in the future. The only possible value for that
configuration was "1" (because Ironic only have 1 API version) and the
ironic team came to an agreement that setting the API version via
configuration option should not be supported anymore.

Closes-Bug: #1458934
Change-Id: Iebfa06da6811889e11fd4edf15d47ca4e0feea6f