Updated openstack/openstack Project: openstack-dev/devstack  7de80c56e5db1f283a71c0ef9fa2304e7d8e2090
Revert "Move default Keystone API version to v3"

It seems like the fallout from this was not well sorted.
A lot of things aren't working, and there is still vestigial
v2 bits left behind.

This should have come with a much greater warning and some
spot checking of additional services working with this.

This reverts commit b162a1d58cdecfb32847b59bd341e06c26efb1ed.

Change-Id: Ia792b23119c00089542ba08879dca1c29dc80945