Updated openstack/openstack Project: openstack/oslosphinx  7de972255141e462e68ac7a04e25ce2b13bc911a
Correct project source guessing for OpenStack

The links to the project source in all the developer docs are
incorrect. The link for keystone is like

 http://git.openstack.org/cgit/p/openstack/keystone

when it should be

 http://git.openstack.org/cgit/openstack/keystone

See the link to "Project Source" on the left on
http://docs.openstack.org/developer/keystone/ for example.

Also, developers sometimes use a ".git" URL when they clone so
strip off the .git.

Change-Id: I8e7620119a169e3322d602ea633303cf9f941605