Updated openstack/openstack Project: openstack/requirements  ef78a41675e1d1e304ca8f92ef3474c5280f8f15
remove argparse from requirements list

argparse is only needed as an external dependency for python 2.6, which
we no longer support. Including it seems to confuse pip 8.0, causing job
failures, so remove it here as a step in removing it from all of the
explicit lists of dependencies in all projects.

Change-Id: Ia57b9a341c977dece684f21f5fc458471a74e7f3
Signed-off-by: Doug Hellmann <doug@doughellmann.com>