Updated openstack/openstack Project: openstack/manila  589c429616a839646d9c350465d554a9b100a8b1
extra-specs should work with string True/False

For a share type extra-spec to match a Boolean capability
(using CapabilitiesFilter), the current design requires
the extra-spec value to use the syntax '<is> True' instead
of just True or 'True'.
When normal humans put 'True' in there, they will find a logged message
from the scheduler saying True is not equal True.

This is because there is a mismatch between the the extra-spec value(string)
and the actual boolean reported.

Fixed the bug by using strutils boolean to string converter.

Change-Id: I959906bc92a374abc25625d012cfbad8aaf11065
Closes-Bug: #1493971