Updated openstack/openstack Project: openstack/nova  cab126a3421b53577e9945a3e010828511ad4191
Spread allocations of fixed ips

The default fixed_ip allocation always does .first() out of the fixed
ip pool. This leads to a pathological situation that in a rapidly
changing environment we're constantly recycling the same small number
of ips out of the pool. Given other racey behavior in dnsmasq around
processing DHCP Release requests, we can start to get DHCP fails when
a DHCP release is missed, we immediately recycle the IP and give it to
a new guest.

We can mitigate this by sorting the results by updated_at. This
means we'll favor least recently used fixed ips.

Depends-On: I27f73c1edf12218818c4d279efbd9fef5cdef672

Related-Bug: #1532809

Change-Id: I3c83bd68a0e2bbbcdd6d955722dbc9f9fc528113