Updated openstack/openstack Project: openstack/ironic  d7f94a9ffcaee37b33a58b4c052211857faceaa6
Enable doc8 style checker for *.rst files

This patch is enabling the doc8 style checker for the *.rst files in
Ironic.

Since this patch is just the first stab at enabling this new check two
errors are currently being ignored by doc8 because of the number of
occurrences in our docs:

* D000 - Duplicated target names. We do have many links with generic
         names scattered around the docs (e.g "see", "link")

* D001 - Line's too long

These errors will be fixed in a following patch.

Change-Id: I32fb5a3104b9773c55a80a84dd81e49c567c76ac