Updated openstack/openstack Project: openstack/nova  21f526f57e8f28fcb9f33bb45cba4478ba2ccb6a
Return HTTP 400 if volume size is not defined

Nova API properly checks that in case of booting from a image to
volume bdm the client shall specify the size of the created volume.
However the API does not handle the exception properly and returns
HTTP 500. This patch handles the InvalidBDM exception and returns
HTTP 400.

Change-Id: I80d0aa64885f25081dad91b1f45da5b318d8ae25
Closes-bug: #1538197