Updated openstack/openstack Project: openstack/nova  0b3b32df4cd75ce5d78646231c0ccee8e94f66ee
Apply scheduler limits to Exact* filters

The DiskFilter, RamFilter and CoreFilter scheduler filters set the
resource limit on the HostState object during filtering. This is
later checked in compute during the claim, and used to enforce
the allocation ratios in a safe manner. The Exact* filters do not
set the resource limits, so scheduler race conditions can result
in multiple bare metal instances claiming a single host.

This change sets the resource limits on the HostState object in
the ExactCoreFilter, ExactDiskFilter and ExactRamFilter, ensuring
that only a single baremetal instance can claim a compute host.

Change-Id: I31d0331afc4698046a4568935a95f70f30e335dd
Partial-Bug: #1341420
Co-Authored-By: Will Miller <wmiller@cray.com>